+++ scraping/src/main/java/com/raccoon/scraper/lastfm/LastfmScraper.java
@@ -19,7 +19,6 @@ import java.util.Set;
 import jakarta.enterprise.context.ApplicationScoped;
 import jakarta.inject.Inject;
 import lombok.extern.slf4j.Slf4j;
-import lombok.val;
 
 @Slf4j
 @ApplicationScoped
@@ -64,26 +63,10 @@ public class LastfmScraper implements TasteScraper {
             log.debug("Processing lastfm artist: {}", lastfmArtist.getName());
             var artist = artistFactory.getOrCreateArtist(lastfmArtist.getName());
             artist.setLastfmUri(lastfmArtist.getUrl());
+            artist.addFollower();
             return artist;
         }
         throw new IllegalArgumentException("Got an object type that is not supported.");
     }
 
-    /**
-     * Guarantees uniqueness of artists per name.
-     * @param base collection to add to
-     * @param from collection to add from
-     * @param seenNames already processed names
-     */
-    private void mergeArtists(final Set<de.umass.lastfm.Artist> base,
-                              final Collection<de.umass.lastfm.Artist> from,
-                              final Set<String> seenNames) {
-        for (val artist : from) {
-            if (seenNames.contains(artist.getName())) {
-                continue;
-            }
-            base.add(artist);
-            seenNames.add(artist.getName());
-        }
-    }
 }